Renamed proto/abstract_proto to proto/abstract
- Fixed build failure in game/world/overlay_elementTests.cpp
