+++ include/proto/abstract/entity.h
@@ -1,7 +1,7 @@
 // This file is subject to the terms and conditions defined in 'LICENSE' in the source code package
 
-#ifndef JACTORIO_INCLUDE_PROTO_ABSTRACT_PROTO_ENTITY_H
-#define JACTORIO_INCLUDE_PROTO_ABSTRACT_PROTO_ENTITY_H
+#ifndef JACTORIO_INCLUDE_PROTO_ABSTRACT_ENTITY_H
+#define JACTORIO_INCLUDE_PROTO_ABSTRACT_ENTITY_H
 #pragma once
 
 #include "jactorio.h"
@@ -154,4 +154,4 @@ namespace jactorio::proto
     };
 } // namespace jactorio::proto
 
-#endif // JACTORIO_INCLUDE_PROTO_ABSTRACT_PROTO_ENTITY_H
+#endif // JACTORIO_INCLUDE_PROTO_ABSTRACT_ENTITY_H
