+++ include/proto/enemy_entity.h
@@ -3,7 +3,7 @@
 #ifndef JACTORIO_INCLUDE_PROTO_ENTITY_ENEMY_ENTITY_H
 #define JACTORIO_INCLUDE_PROTO_ENTITY_ENEMY_ENTITY_H
 
-#include "proto/abstract_proto/entity.h"
+#include "proto/abstract/entity.h"
 
 namespace jactorio::proto
 {
