+++ src/proto/abstract/entity.cpp
@@ -1,6 +1,6 @@
 // This file is subject to the terms and conditions defined in 'LICENSE' included in the source code package
 
-#include "proto/abstract_proto/entity.h"
+#include "proto/abstract/entity.h"
 
 #include "proto/item.h"
