+++ src/render/gui/imgui_manager.cpp
@@ -10,7 +10,7 @@
 
 #include "core/execution_timer.h"
 
-#include "proto/abstract_proto/entity.h"
+#include "proto/abstract/entity.h"
 
 #include "game/player/player_data.h"
 #include "game/world/chunk_tile_layer.h"
