+++ test/jactorioTests.h
@@ -9,7 +9,7 @@
 #include "core/data_type.h"
 
 #include "data/prototype_manager.h"
-#include "proto/abstract_proto/transport_line.h"
+#include "proto/abstract/transport_line.h"
 #include "proto/assembly_machine.h"
 #include "proto/container_entity.h"
 #include "proto/inserter.h"
